requirements.txt
gunicorn==20.1.0
Flask==2.2.2
matplotlib==3.6.3
numpy==1.24.2
opencv-python==4.7.0.68
requests==2.28.2
Pillow==9.4.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
torchvision==0.14.1